Add instructions to run the database image before setting up the application
The database needs some
time to start accepting requests. A fresh building will
fetch the dependencies, and it takes longer than the database
time to start. In scenarios where the dependencies are already
present, there is no time to the database start running, and it
might lead to a failure on application setup.
Running the database container as soon as possible
ensures we will have the database available for `mix setup`.